refactor: atoms 컴포넌트에서 ChallengeDeleteModal 분리
